# 10x speed increase loading miners