I've been profiling the code and seeing where it is slow.  With all miners and algorithms enabled, it takes one of my systems between 5 and 10 minutes to finish ONE loop.

This is the biggest performance increase for the smallest amount of work I've found so far:  Caching Algorithms.txt so it doesn't get read and converted from JSON every time it is called.  The first time Get-Algorithms is called, it will be converted from json and saved to a module local $Algorithms variable.  If it's called again, it checks if that variable is already set and uses the previous results.

## Performance testing

### System:  Windows 10, i7-6800k, SSD, AMD and NVIDIA cards

| Task | Before | After | Percent Increase |
|---|---|---|---|
|Loading pools|27859ms|17798ms|56.5%|
|Loading miners|35709ms|4130ms|764.6%|

### System: Server 2012R2, AMD A10-5800K, regular slow hard drive, NVIDIA card only
| Task | Before | After | Percent Increase|
|---|---|---|---|
|Loading pools|85668ms|40571ms|111.2%|
|Loading miners|133541ms|9615ms|**1288.9%**|