@UselessGuru 2018-05-13
Nice find!

- To further increase poerformance we should to the same trick for Get-Region.
- Most miner files make about at lease 6 Get-Algorithm calls. This can be change to just one call ($Algorithm_Norm = Get-Algorithm $_)

Over Time I'll change the code base of the miners, I'll do this when I add the miner fees feature.