@grantemsley 2018-05-13
@UselessGuru You're right, I missed changing that line to $Script: when I copied it to put in a PR.  Thanks for fixing it :)